+++ connection.go
@@ -457,20 +457,7 @@ func connect(ctx context.Context, c *connection) error {
 		c.pool = conn.NewPool(c.config)
 	}
 
-	c.balancer, err = balancer.New(ctx,
-		c.config, c.pool,
-		append(
-			// prepend common params from root config
-			[]discoveryConfig.Option{
-				discoveryConfig.With(c.config.Common),
-				discoveryConfig.WithEndpoint(c.Endpoint()),
-				discoveryConfig.WithDatabase(c.Name()),
-				discoveryConfig.WithSecure(c.Secure()),
-				discoveryConfig.WithMeta(c.config.Meta()),
-			},
-			c.discoveryOptions...,
-		)...,
-	)
+	c.balancer, err = balancer.New(ctx, c.config, c.pool, c.discoveryOptions...)
 	if err != nil {
 		return xerrors.WithStackTrace(err)
 	}